SML: Introduce `USMLMapAreaTexture` subclass
Create a subclass of `UFGMapAreaTexture` which works in the modding
environment. The intended usage of `USMLMapAreaTexture` is for mods
to create a blueprint child class of it, then set the texture param
to a B8G8R8A8 `Texture2D` that uses one color per map area, and the
editor will automagically fill in the list of areas and their color
from the supplied texture. One should then assign the areas to each
region (`mColorToArea` array element).

When changing the texture, the code will preserve regions that have
a color present in both the old and new textures, and create/delete
the other regions accordingly. However, when unsetting the texture,
regions will be preserved. This is done so that one can refresh the
texture by unsetting and setting it again, given that modifications
to the texture do not propagate to the `USMLMapAreaTexture`.

Should the supplied texture not be valid (its dimensions need to be
square and pixel format needs to be B8G8R8A8), the editor will show
a "fail" notification indicating the reason, and the asset will not
be changed. Editor undo functionality is not tested.

This class will be used in ExampleMod to implement map areas, which
should enable using things like the music manager and the proximity
subsystem in custom maps. This is done in a follow-up.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>