ExampleLevel: Improve map-related functionality
Implement map areas in ExampleLevel using the previously-added map area
texture helper. Also override the minimap texture using a blueprint SCS
hook, as we now have a (rather basic but functional) minimap texture to
show instead of the default map.

Also, properly configure the minimap capture actor so that the position
of things on the minimap accurately reflects the position on the world.
This wasn't working properly because the capture actor not was pointing
downwards, but sideways.

In addition, automatic music is half-working now. Some music plays, but
only "intermission" soundtracks. Biome-specific soundtracks do not play
at all for some reason. It is very likely to be because custom map area
names do not appear in a data table (whose name I don't recall).

Signed-off-by: Angel Pons <th3fanbus@gmail.com>