fix: Make Panini projection usable
The main problem with the Panini projection material function is that
the two inputs to a Lerp node were swapped. Besides fixing that, also
set the default value for Scale to 1 (which matches the default value
in-game) and add a comment to the MPC_Panini parameters about crashes
and a possible (even if miserable) workaround: use the cooked asset.

Tested using a custom material with a Xeno-Basher skeletal mesh and a
custom mesh (red solo cup), equipment appears in the Pioneer's hand.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>