Harden AbstractInstanceManager globals for PIE
At least in the modding environment, trying to use abstract instances
would crash the editor quite often, and compiling blueprints that use
abstract instances would always crash the editor.

Looks like the crashes happen because `StaticManager_PIE` is a global
variable that stores pointers to UObjects, but is not an `UPROPERTY`.
When the GC kills those UObjects, the pointers in the global variable
become dangling pointers without warning.

Make abstract instance managers clean up global state when destroyed,
and use `IsValid()` to avoid returning UObjects that are about to get
killed (as this will result in unpredictable undefined behaviour).

Tested in PIE in the modding environment, this allows seeing abstract
instances in the editor previews (blueprint and level editor) as well
as in PIE worlds. No crashes observed.

FIXME: It is likely that crimes against UE were committed in here...

TODO: There seems to be a bug somewhere that makes abstract instances
in editor worlds get duplicated for some reason, but it does not seem
to cause issues?

NOTE: This also adapts the non-editor code path accordingly, as these
problems could also occur there. However, these changes have not been
validated.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>