ExampleLevel: Use material instances
Consolidate the two materials used for the landscape and the cubes (the
starting platform and the ramp) into a single material. Then create two
material instances of that material to use in the world.

Add an "Emissive Color" parameter to the common material, as it will be
used in a follow-up to show Lumen in action.

As for the reason why material instances were not used to begin with...
This is a mystery I have absolutely no clue about.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>