Fix problems involving inexistent session settings (#252)
* Log an error when a Session Setting cannot be found

It is very easy to mistype the name of a session setting in the code.
Thus, print an error with the name of the session setting in question
so that it is easier to notice the problem when reading the logs.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>

* SessionSettingsManager: Ensure correct type for fallback values

When trying to retrieve the float value of a session setting that does
not exist, the game crashes because SML asserts on something involving
variant type traits:

    Assertion failed: (Type == TVariantTraits<T>::GetType()) || ((TVariantTraits<T>::GetType() == EVariantTypes::UInt8) && (Type == EVariantTypes::Enum))

This is because the fallback value passed to the `FVariant` constructor
template is a literal `0`, whose type is `int`:

    return GetOptionValue(cvar, FVariant(0)).GetValue<float>();
	                            ~~~~~~~~~^~

Using (integer) literals for template argument deduction is troublesome
because the deduced template type is not visible by default. Given that
requiring callers to cast literals is hideous and error-prone, use some
funny template shenanigans to prevent template argument deduction using
the values passed as function arguments. This requires callers of these
function templates to explicitly specify the type argument `T`.

Yes, this approach is absolutely overkill: changing or casting the zero
literals to the correct type would also fix this problem. However, this
approach is more robust and future-proof: bugs that would cause crashes
at runtime are now caught at compile time, and also prevents these bugs
from reappearing in case someone adds support for more types of session
settings in the future.

Note that the "set" functions are not affected by this bug, as they do
not have a literal fallback value. Instead, they use `FVariant(value)`
where `value` is a function parameter. So, template argument deduction
works as long as the function signature has the correct type (always).

Signed-off-by: Angel Pons <th3fanbus@gmail.com>

* Replace TEnableIf with TIdentity_T

---------

Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Co-authored-by: mircearoata <mircearoatapalade@gmail.com>