FGBuildable.cpp: Improve implementation
Be a bit more paranoid when handling abstract instances to avoid NULL
pointer dereferencing. Checking whether `this` is NULL gets optimised
out by most compilers (a NULL `this` is undefined behaviour), so drop
the checks as they are misleading.

Define the `LogBuilding` log category to log a single error, but also
to allow using some inline functions defined in base game headers. If
the log category is not defined, using said inline functions causes a
build-time error.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>