FGBuildable.cpp: Fix root component mobility
Looking at the game's binaries, the root component of buildables has to
default to static mobility. Otherwise, pre-placed buildings do not work
properly in the editor.

The pre-placed buildings in ExampleLevel should no longer be stuck at 0
(which is inside the starting cube), but the level needs to be resaved.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>