SML/Configuration: Make `FillConfigStruct` non-pure
The `UConfigProperty::FillConfigStruct` function is exposed as a pure
function by default because it is `const`. However, since it does not
return anything, there's no good way to know when it will be executed
as pure functions do not have execution pins.

Explicitly make this function non-pure so that it has execution pins.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>