SML/Reflection: Fix NULL check in custom thunks
The `execDeflectStruct` and `execReflectStruct` custom thunks would
throw a blueprint exception when `StructInfo.Struct` is *not* NULL.
This seems to be a mistake, as the log message says that the passed
struct was NULL and `CheckStructParameter` only returns NULL if the
passed struct parameter is invalid.

Flip the NULL checks to make struct reflection and deflection work.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>