AFGGasPillarCloud: Implement editor-only functions
Provide an implementation for this class' three editor-only functions:

 - GatherNearbyPillarLocations
 - NotifyGasPillarRemovedFromInfluence
 - DebugDrawCurrentPillarLocations

Important: `mProximityPillarWorldLocations` stores the location of the
gas pillar plus its `mEffectHeightOffset` (Z axis). Using the pillar's
location alone will not work for big pillars (with a big effect height
offset), but the tiny pillars will work fine because the offset is not
too big. As of writing, the game tests if the distance squared is less
than 40000 (i.e. less than 200 cm) to see if a pillar matched.

Note that these were created from scratch, without seeing the original
implementation by CSS. There may be bugs and/or problems, but at least
the implementations are good enough to make gas pillar clouds function
in ExampleLevel (adapted in the next commit).

Signed-off-by: Angel Pons <th3fanbus@gmail.com>