FGResearchTree: Do not crash/assert when reflecting
Quoting Mircea:

> The whole point of validation is to safely validate the data, not
> crash if it's invalid.

Replace `CastFieldChecked` with `CastField` to avoid asserting when
performing data validation. Also replace one `check` assertion with
a regular NULL check.

For the sake of readability, move the lazy initialisation variables
to helper functions (placed in an anonymous namespace), which fixes
dereferencing possibly NULL pointers. While at it, comment the code
that gets the research tree node class, as it should not be used as
an example on how to do things.

Finally, skip reflection if the research tree has no nodes as there
is nothing to validate.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>