[WIP] ExampleLevel: Set up map areas and music
Setting up map areas requires the previous commit implementing some
editor-only functionality in FGMapAreaTexture. It seeems that there
is no convenient way to initialise a map area texture from BPs only
so we have to fiddle around to make existing game code take care of
it for us.

As for music, it would seem that setting `M Level Started Event` to
`Play_M_MusicManager` is what makes music start playing. This is an
Audiokinetic audio event and is currently not in the SML repo since
the Asset Toolkit lacks support for them.

Although the proximity subsystem seems to be working now, the music
manager doesn't play any biome-specific music. Perhaps this is done
using a different system (Arch said there are "hundreds of volumes"
in the normal map).

Signed-off-by: Angel Pons <th3fanbus@gmail.com>