ExampleMod: Avoid getting the CDO multiple times
The "Get Class Default Object" node is pure, so it will be evaluated
every time its output value is needed. As its return value is stored
in a variable anyway (to stop the CDO from being garbage collected),
we might as well use the value from the variable instead.

Because "Get Class Default Object" prints a warning when accessing a
CDO of a class owned by another plugin, evaluating it twice resulted
in the warning being printed twice. This should no longer happen.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>