ExampleLevel: Improve fallback mesh for resource nodes
When dynamically spawning resource nodes, if the node mesh to use for
a resource was unknown, the code would fall back to the conveyor mesh
and scale it up. However, this can result in very tall resource nodes
which causes problems when trying to place a miner on them.

Instead, scatter some instances of the conveyor mesh around the area.
This uses an instanced static mesh, which will work with any conveyor
mesh (conveyors use instanced static meshes to render items).

Signed-off-by: Angel Pons <th3fanbus@gmail.com>