Change DebugSymbolName from a dynamically-allocated string to a raw string
This is something I wanted to do before, but now I've got an actual
excuse: the next submission will be adding a new handle type, which will
need to store that debug name, and it needs to be as lightweight as
possible.

This technically breaks binary compatibility with vtable and UFunction
hooks, but they haven't reached a release yet so that should be fine. A
backwards-compatible export has been provided for standard hooks.