Add a new handle API for unsubscribing native hooks
This is a replacement for the old UNSUBSCRIBE_* macros, which have now
been removed. Unsubscribing is now a lot easier as the handle will
remember the parameters for you, so you don't need to worry about them
when unsubscribing. This will be necessary (instead of just nice to
have) after a future submission which will add new parameters to virtual
hooks that users can't repeat when unsubscribing.

The signature of UnregisterHook on the backends is now locked in,
instead of letting the backend add custom parameters, because the handle
doesn't know about backend-specific stuff. Fortunately none of the
backends have custom parameters so this isn't an issue.