Fix virtual hooks getting grouped when using different runtime types
Hooks are grouped by the function pointer that they're hooking. However
with virtual functions you can have the same function pointer resolve to
different functions, depending on the SampleObjectInstance that you give
it. This has been worked around by ensuring that each virtual function
hook has its own HookInvoker. This is mildly less efficient if the same
module hooks the same virtual function multiple times, but that's the
best that we can do.