Fix issues when mixing inheritance between SampleObjectInstance and MethodReference
If the MethodReference referred to a base class, but
SampleObjectInstance was of a derived type, then the appropriate
adjustment wasn't being applied to the pointer when finding the vtable.

This has been resolved by changing the SampleObjectInstance parameter to
match the class referred to by MethodReference, instead of just using a
void pointer, so that the adjustment will be done for us.