Fix for hooks getting the wrong 'this' pointer when multiple inheritance is involved
Before we could have cases where a hook thought it had a pointer to a
derived class, when it fact it pointed to a base class, so the pointer
was effectively garbage.

Crimes have been committed to make this work in a way that's backwards
compatible, apologies for the pain and suffering this has caused.